pyproject.toml
[project]
name = "your-project-name"
version = "0.1.0"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
    "asyncpg==0.29.0",
    "aiofiles==23.2.1",
    "python-multipart==0.0.6",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "sqlalchemy==2.0.23",
    "alembic==1.13.0",
]